+++ src/common/routing_table.rs
@@ -1,7 +1,7 @@
 //! Simplified Kademlia routing table
 
+use std::collections::BTreeMap;
 use std::slice::Iter;
-use std::{collections::BTreeMap, convert::TryInto};
 
 use crate::common::{Id, Node, MAX_DISTANCE};
 
@@ -135,58 +135,6 @@ impl RoutingTable {
         nodes
     }
 
-    /// An estimation of the Dht from the distribution of closest nodes
-    /// responding to a query.
-    ///
-    /// In order to get an accurate calculation of the Dht size, you should take
-    /// as many lookups (at uniformly disrtibuted target) as you can,
-    /// and calculate the average of the estimations based on their responding nodes.
-    ///
-    /// # Explanation
-    ///
-    /// Consider a Dht with a 4 bit key space.
-    /// Then we can map nodes in that keyspace by their distance to a given target of a lookup.
-    ///
-    /// Assuming a random but uniform distribution of nodes (which can be measured independently),
-    /// you should see nodes distributed somewhat like this:
-    ///
-    /// ```md
-    ///              (1)    (2)                  (3)    (4)           (5)           (6)           (7)      (8)       
-    /// |------|------|------|------|------|------|------|------|------|------|------|------|------|------|------|
-    /// 0      1      2      3      4      5      6      7      8      9      10     11     12     13     14     15
-    /// ```
-    ///
-    /// So if you make a lookup and optained this partial view of the network:
-    /// ```md
-    ///              (1)    (2)                  (3)                                (4)                  (5)       
-    /// |------|------|------|------|------|------|------|------|------|------|------|------|------|------|------|
-    /// 0      1      2      3      4      5      6      7      8      9      10     11     12     13     14     15
-    /// ```
-    ///
-    /// Note: you see exponentially less further nodes than closer ones, which is what you should expect from how
-    /// the routing table works.
-    ///
-    /// Seeing one node at distance (d1=2), suggests that the routing table might contain 8 nodes,
-    /// since its full length is 8 times (d1).
-    ///
-    /// Similarily, seeing two nodes at (d2=3), suggests that the routing table might contain ~11
-    /// nodes, since the key space is more than (d2).
-    ///
-    /// If we repeat this estimation for as many nodes as the routing table's `k` bucket size,
-    /// and take their average, we get a more accurate estimation of the dht.
-    ///
-    /// ## Formula
-    ///
-    /// The estimated number of Dht size, at each distance `di`, is `en_i = i * d_max / di` where `i` is the
-    /// count of nodes discovered until this distance and `d_max` is the size of the key space.
-    ///
-    /// The final Dht size estimation is the average of `en_1 + en_2 + .. + en_n`
-    ///
-    /// Read more at [A New Method for Estimating P2P Network Size](https://eli.sohl.com/2020/06/05/dht-size-estimation.html#fnref:query-count)
-    pub fn estimate_dht_size(&self) -> usize {
-        estimate_dht_size(self.id, &self.to_vec())
-    }
-
     // === Private Methods ===
 
     #[cfg(test)]
@@ -200,30 +148,6 @@ impl RoutingTable {
         }
         false
     }
-}
-
-pub(crate) fn estimate_dht_size(target: Id, nodes: &[Node]) -> usize {
-    if nodes.is_empty() {
-        return 0;
-    };
-
-    let mut sum = 0;
-    let mut i: usize = 0;
-
-    for node in nodes {
-        let xor = target.xor(&node.id);
-
-        // Round up the lower 4 bytes to get a u128 from u160.
-        let di = u128::from_be_bytes(xor.as_bytes()[0..16].try_into().expect("infallible")) + 1;
-
-        // The inverse of the probability of finding (i) nodes at distance (di)
-        let estimated_n = i.saturating_mul((u128::MAX / di) as usize);
-
-        i += 1;
-        sum += estimated_n;
-    }
-
-    sum / i
 }
 
 impl Default for RoutingTable {